requirements.txt
# Function dependencies
synapseclient>=2.1,<3
google-cloud-secret-manager==1.0.0
